feat: test api endpoints
