+++ src/controllers/tests/pytest.ini
@@ -1,0 +1,8 @@
+[pytest]
+log_cli = true
+log_cli_level = INFO
+log_cli_format = %(asctime)s [%(levelname)s] %(message)s
+log_file = test_result.log
+log_file_level = INFO
+log_file_format = %(asctime)s [%(levelname)s] %(message)s
+log_auto_indent = true
