+++ src/controllers/tests/requirements.txt
@@ -4,11 +4,17 @@ colorama==0.4.6
 Faker==26.0.0
 idna==3.7
 iniconfig==2.0.0
+Jinja2==3.1.4
+MarkupSafe==2.1.5
 packaging==24.1
 pluggy==1.5.0
 pytest==8.3.2
+pytest-html==4.1.1
+pytest-logger==1.1.1
+pytest-metadata==3.1.1
+pytest-sugar==1.0.0
 python-dateutil==2.9.0.post0
 requests==2.32.3
 six==1.16.0
+termcolor==2.4.0
 urllib3==2.2.2
-pprintpp==0.4.0
