+++ src/controllers/tests/test_result.log
@@ -1,0 +1,3 @@
+10:30:11 [INFO] Register Response: {'message': 'user created', 'user': {'name': 'Aaron Jackson', 'email': 'allenjohn@example.com', 'profile': {'first_name': 'Aaron', 'last_name': 'Jackson', 'avatarUrl': '', 'phone_number': None, 'id': '94dca9be-df1f-4973-95fa-31ae6f59b732'}, 'isverified': True, 'google_id': None, 'passwordResetToken': None, 'passwordResetExpires': None, 'id': 'eeadb4ac-7f2c-4bfb-b04b-02d9dfc6e5f2', 'role': 'user', 'createdAt': '2024-08-01T09:30:12.836Z', 'updatedAt': '2024-08-01T09:30:12.836Z', 'is_deleted': False, 'deletedAt': None}, 'access_token': 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJlZWFkYjRhYy03ZjJjLTRiZmItYjA0Yi0wMmQ5ZGZjNmU1ZjIiLCJpYXQiOjE3MjI1MDQ2MTIsImV4cCI6MTcyMjU5MTAxMn0.5ejtwlR0CQ90Pt8IkMia6odMLSMGV9dCWvXMQTCjkeo'}
+10:30:15 [INFO] Register Response: {'message': 'user created', 'user': {'name': 'Emma Green', 'email': 'todd95@example.com', 'profile': {'first_name': 'Emma', 'last_name': 'Green', 'avatarUrl': '', 'phone_number': None, 'id': '28356f1b-0607-45ea-8336-b2de2170f89c'}, 'isverified': True, 'google_id': None, 'passwordResetToken': None, 'passwordResetExpires': None, 'id': '1985f8b6-ea82-46da-9dbb-8ef5dd06b0b9', 'role': 'user', 'createdAt': '2024-08-01T09:30:15.564Z', 'updatedAt': '2024-08-01T09:30:15.564Z', 'is_deleted': False, 'deletedAt': None}, 'access_token': 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxOTg1ZjhiNi1lYTgyLTQ2ZGEtOWRiYi04ZWY1ZGQwNmIwYjkiLCJpYXQiOjE3MjI1MDQ2MTUsImV4cCI6MTcyMjU5MTAxNX0.AzH6WR7AaoZ8eDfyg6mO3dLkMGja19vEwZ8jyT3mMkQ'}
+10:30:20 [INFO] Help Center Topics Response: {'success': True, 'message': 'Fetch Successful', 'data': [], 'status_code': 200}
