+++ src/controllers/tests/tests/run_tests.py
@@ -1,19 +1,16 @@
 import subprocess
 import sys
-import pprintpp
 
 def run_tests():
     python_executable = sys.executable
-    test_command = [python_executable, "-m", "pytest", "-v"]
+    test_command = [python_executable, "-m", "pytest", "--html=report.html", "--self-contained-html", "tests/test_api.py"]
     result = subprocess.run(test_command, capture_output=True, text=True)
     return result.stdout, result.returncode
 
 if __name__ == "__main__":
     output, exit_code = run_tests()
-    # print(output)
-    pprintpp.pprint(output)
+    print(output)
     if exit_code != 0:
         print("Some tests failed.")
     else:
         print("All tests passed.")
-
